[Outliner] Add debug-info support in IR, DWARF and LLDB
This patch addresses the lack of adequate support for outlining in LLVM on three different levels:

        1) In files containing IR code
        2) In DWARF Debugging Information Format
        3) In LLDB debugging tool

Firstly, in IR (.ll) files, outlined functions are currently represented in debugging infromation as
DISubprogram nodes with DIFlagArtificial attribute, which indicates that they are generated by the
compiler. Instead of having such a general characterization, which in no way suggests that a
function is outlined, this patch proposes the use of a new flag for such cases, called
DIFlagOutlined, which is here used in both IROutliner and MachineOutliner optimization passes.

Secondly, since the information that a function has been outlined is now available on IR level, it
can be propagated further through the pipeline and used to differentiate a regular function from
an outlined one, by introducing a new DWARF attribute for the DW_TAG_subprogram DWARF tag, called
DW_AT_LLVM_outlined.

Finally, while debugging an executable using the LLDB debugger, there is currently no proper way of
stating that a function is outlined, outside of the general message given when the address is
compiler-generated code. To fix this, here is proposed the approach of showing a note that the
function, which is being debugged at that time, is outlined. This note is to be written after every
instruction in the outlined function, after printing the matching source code (which in this case
does not correspond to the current function, as it is compiler-generated).